Too many updates....

+++ src/Public/InternetExplorerESC.ps1
@@ -1,8 +1,15 @@
 function Get-InternetExplorerESC {
     $AdminKey = "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components\{A509B1A7-37EF-4b3f-8CFC-4F3A74704073}"
     $UserKey = "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components\{A509B1A8-37EF-4b3f-8CFC-4F3A74704073}"
-    Return @{ Admin = (Get-ItemProperty -Path $AdminKey).IsInstalled
-        User        = (Get-ItemProperty -Path $UserKey).IsInstalled
+    if (Test-Path -path $AdminKey) {
+        Return @{ Admin = (Get-ItemProperty -Path $AdminKey).IsInstalled
+            User        = (Get-ItemProperty -Path $UserKey).IsInstalled
+        }
+    }
+    else {
+        Return @{ Admin = 0
+            User        = 0
+        }
     }
 }
 
@@ -22,11 +29,13 @@ function Set-InternetExplorerESC {
     )
     $AdminKey = "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components\{A509B1A7-37EF-4b3f-8CFC-4F3A74704073}"
     $UserKey = "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components\{A509B1A8-37EF-4b3f-8CFC-4F3A74704073}"
-    switch ($PsCmdlet.ParameterSetName) {
-        'Disable' { if ($DisableAll) {$Admin=0;$User=0} else {$Admin=1;$User=1}  }
-        'Enable' { if ($EnableAll) {$Admin=1;$User=1} else {$Admin=0;$User=0} }
+    if (Test-Path -path $AdminKey) {
+        switch ($PsCmdlet.ParameterSetName) {
+            'Disable' { if ($DisableAll) { $Admin = 0; $User = 0 } else { $Admin = 1; $User = 1 } }
+            'Enable' { if ($EnableAll) { $Admin = 1; $User = 1 } else { $Admin = 0; $User = 0 } }
+        }
+        Set-ItemProperty -Path $AdminKey -Name "IsInstalled" -Value $Admin
+        Set-ItemProperty -Path $UserKey -Name "IsInstalled" -Value $User
+        Stop-Process -Name Explorer
     }
-    Set-ItemProperty -Path $AdminKey -Name "IsInstalled" -Value $Admin
-    Set-ItemProperty -Path $UserKey -Name "IsInstalled" -Value $User
-    Stop-Process -Name Explorer
 }
